Ignore dirty trees when deploying to production
If we make changes to `package.json` and forget to run `npm install` and
commit the resulting `package-lock.json`, then the production working
tree will become dirty when installing because of the `npm install
--production` step.

(This could also happen if the `package-lock.json` file differs after
running `npm install` on one machine vs. another. In general this
_shouldn't_ happen, but I have seen it before, and I haven't looked in
this case so can't rule it out.)

In that case, the _next_ deployment would fail, before this commit
anyway, because `git pull` would refuse to merge changes into a dirty
working tree. So instead of just doing a `git pull`, do a `git fetch
origin` followed by a `git reset --hard origin/master`, which should
always work. This implies that any actual changes made in the production
environment will get blown away, which is in general how deployments
should work anyway: if we do a hotfix to the production environment,
that hotfix should be also added to a commit that goes through the
pipeline and gets deployed in the usual way.