Remove extraneous fontawesome files
I didn't like all those fontawesome files in the repo, cluttering up my
search results and whatnot.

I tried various approaches to simplifying this. There's a
@fortawesome/fontawesome-free NPM package that seems simple, but I'm not
sure how to reference the CSS file within the package. The example here:
https://fontawesome.com/how-to-use/on-the-web/setup/using-package-managers
gives an example that just says the equivalent of
/your-path-to-fontawesome/css/all.css. Presumably I could find this in
the node_modules directory, but then I'd need to mess with the app
configuration to tell it that it can serve files from there, and upon
realizing that, I threw up my hands.

I also tried a "kit", which seems to be the place that the FontAwesome
site steers users to these days. I created an account on fontawesome.com
and added the script tag for my kit, like so:

<script src="https://kit.fontawesome.com/e0b109d7ce.js" crossorigin="anonymous"></script>

But then I ran into issues related to the content security policy we've
added for the site. I fixed the immediate issue by replacing the
existing `helmet.contentSecurityPolicy()` middleware with this:

    app.use(
      helmet.contentSecurityPolicy({
        useDefaults: true,
        directives: {
          defaultSrc: ["'self'", 'kit.fontawesome.com'],
        },
      })
    )

But then I was having issues with the second level of requests, which
were going to something like ka-f.fontawesome.com/something-or-other. I
didn't see anything in the documentation about that, including in the
page that specifically talks about CSP stuff:
https://fontawesome.com/how-to-use/on-the-web/other-topics/security
...so I wasn't sure whether that origin was an implementation detail,
subject to change later and break things if I hardcoded the CSP to
include it, or what. So I gave up on that approach too.

This led me back to the idea of just vendoring the necessary files
in-repo. I realized that most of what we vendored wasn't ever going to
be used and could be safely deleted. With some trial and error, I
learned that we need not only the "all" css file but also the web fonts
themselves. Specifically, the `fa-solid-900.woff2` file seems to be the
only one the app is actually requesting. Part of that is because all of
the current uses of FA are with the `fas` style, meaning `fa-solid`, but
I'm not sure whether the `-900` or `.woff2` fragments are based on our
usage, my current browser, or what. So I just left all the web fonts in.
There are only about a dozen, which doesn't clutter things up too much.