Fix header detection of spreadsheet uploads
Before, we were actually testing the first row of data, which might
miss some non-required fields. That would produce a simple
'incorrect headers' error message, which wasn't helpful.

Now, a better error message is used, and the actual headers of the
worksheet are used to compare against the expected fields.