Avoid a failed test by hardcoding values
For some reason the student spec was failing, but only when run as
part of the whole test suite. In that case, it failed on the
beforeAll function because the body it got back from the request was
null/undefined.

I'm out of my depth with Cypress, especially as it pertains to
handling asynchrony. I suspect a race condition of some kind.

I tried appending `.debug()` to the `cy.request` call, which let me
poke around a bit. I discovered that the response apparently doesn't
have a content length header, and the body is apparently empty.

I can't figure out how that could be. The response code is still
200, so it's not like there was an error, because then the code
would be 500. I tried printing the to-be-returned roles from the
fixture module, and they were reasonable every time.

Anyway, I eventually realized I could just couple the spec to the
fixture data values. Not ideal, but I don't have time to fix this
right now.