Fix problems with asynchrony in student upload handler
We were calling `res.render()` repeatedly when there were validation
errors with multiple lines of the input. This was causing problems
in Chrome and Edge, but strangely not Firefox.

The solution is to check for various kinds of errors first, and only
call `res.render()` to render an error page if there was an error,
and then only for the first error rather than each of them.